# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Comment:
    properties:
      content:
        example: This is a comment
        type: string
      id:
        type: integer
      postId:
        example: 1
        type: integer
      reactionInfo:
        $ref: '#/definitions/api.ReactionInfo'
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
...
  api.ErrorDetails:
    properties:
      code:
        type: integer
      details:
        items:
          $ref: '#/definitions/api.ValidationError'
        type: array
      message:
        type: string
    type: object
...
  api.GroupResponseInfo:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  api.Metadata:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
...
  api.Post:
    properties:
      commentCount:
        type: integer
      content:
        example: Test content
        type: string
      group:
        $ref: '#/definitions/api.GroupResponseInfo'
      id:
        type: integer
      image:
        type: string
      privacy:
        example: public
        type: string
      reactionInfo:
        $ref: '#/definitions/api.ReactionInfo'
      title:
        example: Test title
        type: string
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
  api.PostCreateRequest:
    properties:
      content:
        example: Test content
        type: string
      group:
        $ref: '#/definitions/api.GroupResponseInfo'
      image:
        type: string
      privacy:
        example: public
        type: string
      title:
        example: Test title
        type: string
    type: object
  api.PostResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/api.Comment'
        type: array
      post:
        $ref: '#/definitions/api.Post'
    type: object
  api.PostsListResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/api.Post'
        type: array
    type: object
...
  api.ReactionInfo:
    properties:
      dislikeCount:
        type: integer
      likeCount:
        type: integer
      status:
        type: string
    type: object
...
  api.Response:
    properties:
      authenticated:
        type: boolean
      message:
        type: string
      metadata:
        $ref: '#/definitions/api.Metadata'
      status:
        type: string
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
  api.ResponseID:
    properties:
      id:
        type: integer
    type: object
  api.User:
    properties:
      aboutMe:
        type: string
      avatar:
        type: string
      bannerColor:
        type: string
      commentCount:
        type: integer
      createdAt:
        type: string
      dateOfBirth:
        type: string
      firstName:
        type: string
      followerCount:
        type: integer
      followingCount:
        type: integer
      id:
        type: integer
      lastActive:
        type: string
      lastName:
        type: string
      nickname:
        type: string
      postCount:
        type: integer
      profileVisibility:
        type: string
    type: object
  api.UserPage:
    properties:
      chatHash:
        type: string
      followStatus:
        type: boolean
      followers:
        items:
          $ref: '#/definitions/api.UserResponseInfo'
        type: array
      following:
        items:
          $ref: '#/definitions/api.UserResponseInfo'
        type: array
      groupsPosts:
        items:
          $ref: '#/definitions/api.Post'
        type: array
      personalPosts:
        items:
          $ref: '#/definitions/api.Post'
        type: array
      user:
        $ref: '#/definitions/api.User'
    type: object
  api.UserResponseInfo:
    properties:
      id:
        type: integer
      nickname:
        type: string
    type: object
  api.UsersList:
    properties:
      users:
        items:
          $ref: '#/definitions/api.User'
        type: array
    type: object
  api.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for kood-social-network.
  title: kood-social-network API
  version: 0.1.0
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Retrieve all public posts and those from users who have accepted
        follow requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.PostsListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Get all posts for an authenticated user
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Creates a new post with the provided post details and optional
        image file.
      parameters:
      - description: Post details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PostCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.ResponseID'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Get a post by ID
      tags:
      - posts
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.UsersList'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Get all users
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.UserPage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Get user by ID
      tags:
      - users
swagger: "2.0"